django/{{project_name}}/{{project_name}}/apps/core/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status


@api_view(['GET'])
def health_check(request):
    """Health check endpoint."""
    return Response({
        'status': 'healthy',
        'service': '{{project_name}}',
    }, status=status.HTTP_200_OK)